refactor upload process to be run as a resque task. to alleviate stress on the server when a user is uploading a script.
