+++ app/views/upload/uploadfile.html.erb
@@ -28,7 +28,7 @@
         <%= content_tag :div, :class=>"form-group" do %>
              <%= label_tag("Select Characters Per Line to Display") %>
           </br>
-          <%= select_tag "character_per_line", options_for_select(20..50, 40) , :class=>"form-control" %>
+          <%= select_tag "characters_per_line", options_for_select(20..50, 40) , :class=>"form-control" %>
         <% end %>
 
         <%= content_tag :div, :class=>"form-group" do %>
